cleanups
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@17869 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ willow/src/wgetstats/wgetstats.cc
@@ -206,7 +206,7 @@ uint8_t		*vers;
 		bufp += 4;
 		hi = *(uint32_t *) bufp;		
 		bufp += 4;
-		nc = hi << 32 | lo;
+		nc = (uint64_t)hi << 32 | lo;
 		fprintf(stderr, "\tListener %.*s: %llu connections\n",
 			nlen, name, nc);
 	}
